Refactored to Abstract* classes to allow customizing.
