+++ spring-cloud-config-client/src/main/java/org/springframework/cloud/config/client/ConfigServerConfigDataLocation.java
@@ -16,74 +16,14 @@
 
 package org.springframework.cloud.config.client;
 
-import java.util.Objects;
-
-import org.apache.commons.logging.Log;
-
-import org.springframework.boot.context.config.ConfigDataLocation;
 import org.springframework.boot.context.config.Profiles;
-import org.springframework.core.style.ToStringCreator;
 import org.springframework.web.client.RestTemplate;
 
-public class ConfigServerConfigDataLocation extends ConfigDataLocation {
+public class ConfigServerConfigDataLocation extends AbstractConfigDataLocation {
 
-	private final Log log;
-
-	private final RestTemplate restTemplate;
-
-	private final ConfigClientProperties properties;
-
-	private final Profiles profiles;
-
-	public ConfigServerConfigDataLocation(Log log, RestTemplate restTemplate,
+	public ConfigServerConfigDataLocation(RestTemplate restTemplate,
 			ConfigClientProperties properties, Profiles profiles) {
-		this.log = log;
-		this.restTemplate = restTemplate;
-		this.properties = properties;
-		this.profiles = profiles;
-	}
-
-	public Log getLog() {
-		return this.log;
-	}
-
-	public RestTemplate getRestTemplate() {
-		return this.restTemplate;
-	}
-
-	public ConfigClientProperties getProperties() {
-		return this.properties;
-	}
-
-	public Profiles getProfiles() {
-		return this.profiles;
-	}
-
-	@Override
-	public boolean equals(Object o) {
-		if (this == o) {
-			return true;
-		}
-		if (o == null || getClass() != o.getClass()) {
-			return false;
-		}
-		ConfigServerConfigDataLocation that = (ConfigServerConfigDataLocation) o;
-		return Objects.equals(this.log, that.log)
-				&& Objects.equals(this.restTemplate, that.restTemplate)
-				&& Objects.equals(this.properties, that.properties)
-				&& Objects.equals(this.profiles, that.profiles);
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(this.log, this.restTemplate, this.properties, this.profiles);
-	}
-
-	@Override
-	public String toString() {
-		return new ToStringCreator(this).append("uris", properties.getUri())
-				.append("profiles", profiles.getAccepted()).toString();
-
+		super(restTemplate, properties, profiles);
 	}
 
 }
