+++ spring-cloud-config-client/src/test/resources/META-INF/spring.factories
@@ -1,0 +1,8 @@
+
+# ConfigData Location Resolvers
+org.springframework.boot.context.config.ConfigDataLocationResolver=\
+org.springframework.cloud.config.client.TestConfigServerConfigDataLocationResolver
+
+# ConfigData Loaders
+org.springframework.boot.context.config.ConfigDataLoader=\
+org.springframework.cloud.config.client.TestConfigServerConfigDataLoader
